requirements.txt
# This file was autogenerated by uv v0.1.5 via the following command:
#    uv pip compile - -o requirements.txt
biopython==1.83
numpy==1.26.4
    # via
    #   biopython
    #   pandas
pandas==2.2.2
python-dateutil==2.9.0.post0
    # via pandas
pytz==2024.1
    # via pandas
six==1.16.0
    # via python-dateutil
tzdata==2024.1
    # via pandas
